requirements.txt
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
Flask==3.1.2
Flask-Session==0.8.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.19.0
passlib==1.7.4
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.13.0
redis==6.4.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
